Rename declarations file
