+++ src/main/modules/config.ts
@@ -6,7 +6,7 @@ import path from 'path';
 import electron from 'electron';
 import teeny from 'teeny-conf';
 
-import { Config, Repeat, SortBy, SortOrder } from '../../shared/types/interfaces';
+import { Config, Repeat, SortBy, SortOrder } from '../../shared/types/museeks';
 import Module from './module';
 
 const { app } = electron;
