+++ src/main/modules/mpris.ts
@@ -6,7 +6,7 @@ import electron from 'electron';
 import * as mpris from 'mpris-service';
 import * as mime from 'mime-types';
 
-import { TrackModel } from '../../shared/types/interfaces';
+import { TrackModel } from '../../shared/types/museeks';
 import { fetchCover } from '../../shared/utils/cover';
 import { SUPPORTED_TRACKS_EXTENSIONS } from '../../shared/constants';
 import ModuleWindow from './module-window';
