+++ src/ui/actions/PlayerActions.ts
@@ -11,7 +11,7 @@ import * as utils from '../lib/utils';
 import Player from '../lib/player';
 import { sortTracks, filterTracks } from '../lib/utils-library';
 import { shuffleTracks } from '../lib/utils-player';
-import { TrackModel, PlayerStatus, Repeat } from '../../shared/types/interfaces';
+import { TrackModel, PlayerStatus, Repeat } from '../../shared/types/museeks';
 import * as ToastsActions from './ToastsActions';
 
 const { ipcRenderer } = electron;
