+++ src/ui/components/Header/Header.tsx
@@ -9,7 +9,7 @@ import PlayerControls from '../PlayerControls/PlayerControls';
 import * as LibraryActions from '../../actions/LibraryActions';
 import { isCtrlKey } from '../../lib/utils-platform';
 import { RootState } from '../../reducers';
-import { TrackModel, PlayerStatus, Repeat } from '../../../shared/types/interfaces';
+import { TrackModel, PlayerStatus, Repeat } from '../../../shared/types/museeks';
 
 import styles from './Header.module.css';
