+++ src/ui/components/PlayerControls/PlayerControls.tsx
@@ -3,7 +3,7 @@ import Icon from 'react-fontawesome';
 import VolumeControl from '../VolumeControl/VolumeControl';
 
 import * as PlayerActions from '../../actions/PlayerActions';
-import { PlayerStatus } from '../../../shared/types/interfaces';
+import { PlayerStatus } from '../../../shared/types/museeks';
 
 import styles from './PlayerControls.module.css';
