+++ src/ui/lib/utils-library.ts
@@ -1,5 +1,5 @@
 import orderBy from 'lodash-es/orderBy';
-import { TrackModel } from '../../shared/types/interfaces';
+import { TrackModel } from '../../shared/types/museeks';
 import * as utils from './utils';
 
 /**
