+++ src/ui/reducers/app.ts
@@ -1,5 +1,5 @@
 import types from '../constants/action-types';
-import { Action } from '../../shared/types/interfaces';
+import { Action } from '../../shared/types/museeks';
 
 export type AppState = null;
