+++ src/ui/reducers/library.ts
@@ -2,7 +2,7 @@ import types from '../constants/action-types';
 
 import { config } from '../lib/app';
 import * as utils from '../lib/utils';
-import { Action, TrackModel, SortBy, SortOrder } from '../../shared/types/interfaces';
+import { Action, TrackModel, SortBy, SortOrder } from '../../shared/types/museeks';
 
 export interface LibrarySort {
   by: SortBy;
