+++ src/ui/reducers/toasts.ts
@@ -1,6 +1,6 @@
 import types from '../constants/action-types';
 
-import { Toast, Action } from '../../shared/types/interfaces';
+import { Toast, Action } from '../../shared/types/museeks';
 
 export type ToastsState = Toast[];
