+++ src/ui/store.ts
@@ -4,7 +4,7 @@ import { createLogger } from 'redux-logger';
 import { persistStore, persistReducer, createTransform } from 'redux-persist';
 import storage from 'redux-persist/lib/storage';
 
-import { PlayerStatus } from '../shared/types/interfaces';
+import { PlayerStatus } from '../shared/types/museeks';
 import rootReducer, { RootState } from './reducers';
 import { PlayerState } from './reducers/player';
 import player from './lib/player';
